Refactor: 깃 이슈 생성 파일 분리 및 주석 추가
