+++ Pick/src/main/java/com/nob/pick/gitactivity/command/application/controller/GitHubAccountController.java
@@ -1,9 +1,7 @@
 package com.nob.pick.gitactivity.command.application.controller;
 
-import com.nob.pick.common.util.JwtUtil;
 import com.nob.pick.gitactivity.command.application.dto.GitHubAccountDTO;
 import com.nob.pick.gitactivity.command.application.service.GitHubAccountService;
-import com.nob.pick.gitactivity.command.application.service.GitHubActivityService;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
@@ -14,24 +12,17 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.io.IOException;
-import java.util.Map;
 
-@RestController
+// 깃에 로그인해서 토큰을 받아오고 저장하는 기능 관리
 @Slf4j
+@RestController
 @RequestMapping("/api/github")
 public class GitHubAccountController {
-    private final JwtUtil jwtUtil;
     private final GitHubAccountService gitHubAccountService;
-    private final GitHubActivityService gitHubActivityService;
 
     @Autowired
-    public GitHubAccountController(
-            JwtUtil jwtUtil
-            , GitHubAccountService gitHubAccountService
-            , GitHubActivityService gitHubActivityService) {
-        this.jwtUtil = jwtUtil;
+    public GitHubAccountController(GitHubAccountService gitHubAccountService) {
         this.gitHubAccountService = gitHubAccountService;
-        this.gitHubActivityService = gitHubActivityService;
     }
 
     // 자동 연동 콜백 처리 (클라이언트가 GitHub 로그인 후 호출)
@@ -87,34 +78,9 @@ public class GitHubAccountController {
         }
     }
 
-    private String extractJwt(HttpServletRequest request) {
-        String header = request.getHeader("Authorization");
-        if (header != null && header.startsWith("Bearer ")) {
-            return header.substring(7);
-        }
-        return null;
-    }
-
     @GetMapping("/success")
     public String successPage() {
         return "GitHub 연동이 완료되었습니다!";
-    }
-
-    // 사용자가 버튼 클릭 → 이슈 생성 API
-    @PostMapping("/issue")
-    public ResponseEntity<?> createIssue(@RequestBody Map<String, String> body, HttpServletRequest request) {
-        String jwt = extractJwt(request);
-        String repo = body.get("repo");
-        String title = body.get("title");
-        String content = body.getOrDefault("body", "");
-
-        int memberId = jwtUtil.getId(jwt);
-        
-        // 🚩 memberId를 통해 member를 찾고 해당 데이터의 githubAccountId 값 가져오기
-        int gitHubAccountId = 1;        // 우선 임의로 1 저장
-
-        gitHubActivityService.createGitIssue(gitHubAccountId, repo, title, content);
-        return ResponseEntity.ok("이슈 생성 완료");
     }
 
 }
